Added ToC, About and Compile Instruction sections
I have added a disclaimer at the top of the README.md to state that the README is still under construction. I also added a table of contents, I think this will make navigating the README in the future easier and it is easier to update. I also included a draft of the about section for the project and included a compile instructions section that will be updated at a later time.